Factor out get-entries command to a file
